% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{hist.ctdf}
\alias{hist.ctdf}
\title{Histogram of observation times by cluster}
\usage{
\method{hist}{ctdf}(x, binwidth = 3600, ...)
}
\arguments{
\item{x}{A \code{ctdf} object.}

\item{binwidth}{Bin width in seconds (POSIXct is binned in seconds). Defaults to 1 hour.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot}.
}
\description{
Plot per-cluster histograms of observation times with cluster start/stop markers,
stacked vertically.
}
\examples{
require(clusterTrack.Vis)
data(pesa56511)
ctdf = as_ctdf(pesa56511, time = "locationDate") |> cluster_track()
hist(ctdf)
}
