% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{site}
\alias{site}
\title{Build a simple static site index for a folder of leaflet/htmlwidgets maps}
\usage{
site(path)
}
\arguments{
\item{path}{Directory containing saved map \code{.html} files.}
}
\value{
Invisibly, a list with paths (\code{index}, \code{assets_dir}, \code{libs_dir},
\code{metadata_dir}) and a \code{data.table} summary of discovered maps.
}
\description{
Creates an \code{index.html} with links and inline previews for all \code{.html} maps
under \code{path}
}
\examples{
\dontrun{
siteloc = "~/Desktop/temp"
data(mini_ruff)
mini_ruff = as_ctdf(mini_ruff) |> cluster_track()
map(mini_ruff, path = siteloc)

data(pesa56511)
pesa = as_ctdf(pesa56511, time = "locationDate") |> cluster_track()
map(pesa, path = siteloc)

data(ruff143789)
ruff = as_ctdf(ruff143789, time = "locationDate") |> cluster_track()
map(ruff, path = siteloc)

data(lbdo66862)
lbdo = as_ctdf(lbdo66862, time = "locationDate") |> cluster_track()
map(lbdo, path = siteloc)

site(siteloc)


}
}
