% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_ctdf}
\alias{summarise_ctdf}
\title{Extended summary of a ctdf}
\usage{
summarise_ctdf(ctdf)
}
\arguments{
\item{ctdf}{A \code{ctdf} object.}
}
\description{
Extends the clusterTrack::summary(ctdf)
}
\examples{
data(pesa56511)
ctdf = as_ctdf(pesa56511, time = "locationDate") |> cluster_track()

summarise_ctdf(ctdf)
}
