% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Map a ctdf object}
\usage{
map(ctdf, path, name, fix_dateline = FALSE)
}
\arguments{
\item{ctdf}{A \code{ctdf} object.}

\item{path}{Optional \code{character}. File path where the HTML map will be saved.
If omitted, the function returns the \code{leaflet} map object.}

\item{name}{When path is given, the name of the html file.}
}
\value{
A \code{leaflet} map object if \code{path} is not provided.
If \code{path} is provided, an HTML file (and its \verb{assets/} directory) is written to disk.
}
\description{
Visualize clustered track data on an interactive map.
}
\examples{
data(mini_ruff)
ctdf = as_ctdf(mini_ruff)
map(ctdf) # empty map
cluster_track(ctdf)
map(ctdf)
map(ctdf, path = "~/Desktop/temp")
}
\seealso{
\code{\link[clusterTrack:cluster_track]{clusterTrack::cluster_track()}}, \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}
}
