% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{site}
\alias{site}
\title{Create a Quarto index template for browsing saved maps}
\usage{
site(path)
}
\arguments{
\item{path}{Directory where the template \code{index.qmd} should be copied.}
}
\description{
Copies a Quarto \code{index.qmd} template into \code{path}. Render the copied file
(for example with Quarto) to generate an \code{index.html} that can be used to
navigate to the saved map \code{.html} files in the folder.
}
\examples{
\dontrun{
siteloc = "~/Desktop/temp"
data(mini_ruff)
mini_ruff = as_ctdf(mini_ruff) |> cluster_track()
map(mini_ruff, path = siteloc)

data(pesa56511)
pesa = as_ctdf(pesa56511, time = "locationDate") |> cluster_track()
map(pesa, path = siteloc)

data(ruff143789)
ruff = as_ctdf(ruff143789, time = "locationDate") |> cluster_track()
map(ruff, path = siteloc)

data(lbdo66862)
lbdo = as_ctdf(lbdo66862, time = "locationDate") |> cluster_track()
map(lbdo, path = siteloc)

site(siteloc)


}
}
